# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Segurança Digital",
        "description": "Documentação do Projeto Sistema de Segurança Digital",
        "contact": {
            "name": "João Paulo Pinheiro",
            "email": "meu@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://segurancadigital.local/api/v1",
            "description": "Ambiente de Desenvolvimento"
        },
        {
            "url": "http://segurancadigital.prod/api/v1",
            "description": "Ambiente de Produção"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Logar no sistema JWT por meio de e-mail e senha.",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "super@super.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Senha",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "super"
                    },
                    {
                        "name": "remember",
                        "in": "query",
                        "description": "Lembre-me",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Normal Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/LoginProperty"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Retorna os dados do usuário logado.",
                "operationId": "App\\Http\\Controllers\\AuthController::me",
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/MeResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/payoad": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Retorna dados de payload JWT.",
                "operationId": "App\\Http\\Controllers\\AuthController::payoad",
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PayloadResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Desconecte o usuário (invalide o token).",
                "operationId": "App\\Http\\Controllers\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Retorno padrão",
                "description": "Entidade de resposta, resultado da resposta usa essa estrutura uniformemente",
                "properties": {
                    "success": {
                        "description": "Status da resposta",
                        "type": "boolean"
                    },
                    "code": {
                        "description": "Código da resposta",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Mensagem da resposta",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ExpirePayloadProperty": {
                "title": "Auth Payload Expire",
                "description": "Datas Payload de Usuário",
                "properties": {
                    "emitido_em": {
                        "description": "emitido_em",
                        "type": "string"
                    },
                    "expira_em": {
                        "description": "expira_em",
                        "type": "string"
                    },
                    "nao_antes_de": {
                        "description": "nao_antes_de",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginProperty": {
                "title": "Auth Login",
                "description": "Response JWT",
                "properties": {
                    "access_token": {
                        "description": "access_token",
                        "type": "string"
                    },
                    "token_type": {
                        "description": "token_type",
                        "type": "string"
                    },
                    "expires_in_minutes": {
                        "description": "expires_in_minutes",
                        "type": "integer"
                    },
                    "expires_in_date": {
                        "description": "expires_in_date",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "MeResponse": {
                "title": "Auth Me",
                "description": "Retorn dados do usuário",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "email": {
                        "description": "email",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "email_verified_at",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string"
                    },
                    "rules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleWithPermissionsProperty"
                        }
                    }
                },
                "type": "object"
            },
            "PayloadProperty": {
                "title": "Auth Payload Property",
                "description": "Property Payload de Usuário",
                "properties": {
                    "iss": {
                        "description": "iss",
                        "type": "string"
                    },
                    "iat": {
                        "description": "iat",
                        "type": "integer"
                    },
                    "exp": {
                        "description": "exp",
                        "type": "integer"
                    },
                    "nbf": {
                        "description": "nbf",
                        "type": "integer"
                    },
                    "jti": {
                        "description": "jti",
                        "type": "string"
                    },
                    "sub": {
                        "description": "sub",
                        "type": "integer"
                    },
                    "prv": {
                        "description": "prv",
                        "type": "string"
                    },
                    "email": {
                        "description": "email",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PayloadResponse": {
                "title": "Auth Payload",
                "description": "Response Payload de Usuário",
                "properties": {
                    "payload": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PayloadProperty"
                        }
                    },
                    "datas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExpirePayloadProperty"
                        }
                    }
                },
                "type": "object"
            },
            "PermissionProperty": {
                "title": "Auth Permissions",
                "description": "Permissões de Usuário",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "display name",
                        "type": "string"
                    },
                    "description": {
                        "description": "description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "RoleWithPermissionsProperty": {
                "title": "Auth Roles/Permissions",
                "description": "Perfil de Usuário e Permissões",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "display name",
                        "type": "string"
                    },
                    "description": {
                        "description": "description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionProperty"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Informe o token JWT para endpoits seguros",
                "name": "JWT bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}